# openapi/openapi.v1.0.0-rc1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StrategyEvaluateRequest": {
        "properties": {
          "analysis_notes": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Analysis Notes",
            "type": "array"
          },
          "balance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance"
          },
          "candle_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candle Count"
          },
          "check_risk": {
            "default": true,
            "title": "Check Risk",
            "type": "boolean"
          },
          "entry_price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Price"
          },
          "equity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional equity override for offline evaluation",
            "title": "Equity"
          },
          "fvg_bearish": {
            "default": false,
            "title": "Fvg Bearish",
            "type": "boolean"
          },
          "fvg_bullish": {
            "default": false,
            "title": "Fvg Bullish",
            "type": "boolean"
          },
          "has_fvgs": {
            "default": false,
            "title": "Has Fvgs",
            "type": "boolean"
          },
          "has_liquidity": {
            "default": false,
            "title": "Has Liquidity",
            "type": "boolean"
          },
          "has_order_blocks": {
            "default": false,
            "title": "Has Order Blocks",
            "type": "boolean"
          },
          "has_structure": {
            "default": false,
            "title": "Has Structure",
            "type": "boolean"
          },
          "last_price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Price"
          },
          "liquidity_sweep_bearish": {
            "default": false,
            "title": "Liquidity Sweep Bearish",
            "type": "boolean"
          },
          "liquidity_sweep_bullish": {
            "default": false,
            "title": "Liquidity Sweep Bullish",
            "type": "boolean"
          },
          "market_open": {
            "default": true,
            "title": "Market Open",
            "type": "boolean"
          },
          "order_block_bearish": {
            "default": false,
            "title": "Order Block Bearish",
            "type": "boolean"
          },
          "order_block_bullish": {
            "default": false,
            "title": "Order Block Bullish",
            "type": "boolean"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "requested_lots": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Lots"
          },
          "session": {
            "default": "unknown",
            "maxLength": 32,
            "title": "Session",
            "type": "string"
          },
          "stop_loss_distance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss Distance"
          },
          "structure_bias": {
            "default": "unknown",
            "description": "up | down | range | unknown (from Market Structure)",
            "title": "Structure Bias",
            "type": "string"
          },
          "symbol": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          },
          "tick_age_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional tick age override (tests / offline)",
            "title": "Tick Age Seconds"
          },
          "timeframe": {
            "default": "m15",
            "maxLength": 16,
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "symbol"
        ],
        "title": "StrategyEvaluateRequest",
        "type": "object"
      },
      "StrategyEvaluateResponse": {
        "properties": {
          "decision": {
            "description": "no_action | watch | ready | blocked",
            "title": "Decision",
            "type": "string"
          },
          "evaluated_at": {
            "format": "date-time",
            "title": "Evaluated At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "market_state": {
            "additionalProperties": true,
            "title": "Market State",
            "type": "object"
          },
          "preconditions": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Preconditions",
            "type": "object"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "risk_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Decision"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "signal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StrategySignalResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "id",
          "request_id",
          "symbol",
          "timeframe",
          "decision",
          "evaluated_at"
        ],
        "title": "StrategyEvaluateResponse",
        "type": "object"
      },
      "StrategySignalListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/StrategySignalResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "StrategySignalListResponse",
        "type": "object"
      },
      "StrategySignalResponse": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "direction": {
            "title": "Direction",
            "type": "string"
          },
          "evaluation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Id"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "rejected": {
            "default": false,
            "title": "Rejected",
            "type": "boolean"
          },
          "rejection_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Rejection Reasons",
            "type": "array"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "id",
          "symbol",
          "timeframe",
          "direction",
          "confidence",
          "generated_at"
        ],
        "title": "StrategySignalResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WalkForwardBarRequest": {
        "properties": {
          "close": {
            "title": "Close",
            "type": "string"
          },
          "close_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Time"
          },
          "high": {
            "title": "High",
            "type": "string"
          },
          "low": {
            "title": "Low",
            "type": "string"
          },
          "open": {
            "title": "Open",
            "type": "string"
          },
          "open_time": {
            "title": "Open Time",
            "type": "string"
          },
          "volume": {
            "default": "0",
            "title": "Volume",
            "type": "string"
          }
        },
        "required": [
          "open_time",
          "open",
          "high",
          "low",
          "close"
        ],
        "title": "WalkForwardBarRequest",
        "type": "object"
      },
      "WalkForwardListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/WalkForwardRunResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "WalkForwardListResponse",
        "type": "object"
      },
      "WalkForwardRunRequest": {
        "properties": {
          "anchored": {
            "default": false,
            "title": "Anchored",
            "type": "boolean"
          },
          "auto_analysis": {
            "default": true,
            "title": "Auto Analysis",
            "type": "boolean"
          },
          "bars": {
            "items": {
              "$ref": "#/components/schemas/WalkForwardBarRequest"
            },
            "title": "Bars",
            "type": "array"
          },
          "in_sample_bars": {
            "default": 40,
            "maximum": 5000.0,
            "minimum": 5.0,
            "title": "In Sample Bars",
            "type": "integer"
          },
          "initial_balance": {
            "default": "10000",
            "title": "Initial Balance",
            "type": "string"
          },
          "optimize_params": {
            "default": true,
            "title": "Optimize Params",
            "type": "boolean"
          },
          "out_of_sample_bars": {
            "default": 20,
            "maximum": 2000.0,
            "minimum": 2.0,
            "title": "Out Of Sample Bars",
            "type": "integer"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "step_bars": {
            "default": 20,
            "maximum": 2000.0,
            "minimum": 1.0,
            "title": "Step Bars",
            "type": "integer"
          },
          "symbol": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "default": "m15",
            "maxLength": 16,
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "symbol"
        ],
        "title": "WalkForwardRunRequest",
        "type": "object"
      },
      "WalkForwardRunResponse": {
        "properties": {
          "aggregated_is": {
            "additionalProperties": true,
            "title": "Aggregated Is",
            "type": "object"
          },
          "aggregated_oos": {
            "additionalProperties": true,
            "title": "Aggregated Oos",
            "type": "object"
          },
          "bar_count": {
            "default": 0,
            "title": "Bar Count",
            "type": "integer"
          },
          "combined_equity": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Combined Equity",
            "type": "array"
          },
          "error_message": {
            "default": "",
            "title": "Error Message",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "fold_count": {
            "default": 0,
            "title": "Fold Count",
            "type": "integer"
          },
          "folds": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Folds",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "promotion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "promote_to_paper | needs_rework | reject",
            "title": "Promotion"
          },
          "report": {
            "additionalProperties": true,
            "title": "Report",
            "type": "object"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "robustness": {
            "additionalProperties": true,
            "title": "Robustness",
            "type": "object"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "title": "Timeframe",
            "type": "string"
          },
          "window_config": {
            "additionalProperties": true,
            "title": "Window Config",
            "type": "object"
          }
        },
        "required": [
          "id",
          "request_id",
          "symbol",
          "timeframe",
          "status"
        ],
        "title": "WalkForwardRunResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "QuantForg v1.0.0 Release Candidate 1 OpenAPI specification. EXECUTION_ENABLED remains false. No AI endpoints. Generated from FastAPI route registrations.",
    "title": "QuantForg API",
    "version": "1.0.0-rc.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/strategy/evaluate": {
      "post": {
        "description": "Run the Strategy Runtime pipeline.\n\nReturns NO_ACTION | WATCH | READY | BLOCKED. Never places an order.\nNever enables EXECUTION_ENABLED. Never calls order_send().",
        "operationId": "strategy_evaluate_api_v1_strategy_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyEvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyEvaluateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Strategy Evaluate",
        "tags": [
          "strategy-runtime"
        ]
      }
    },
    "/api/v1/strategy/signals": {
      "get": {
        "description": "List Strategy Runtime signals for the current user (history only).",
        "operationId": "strategy_signals_api_v1_strategy_signals_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include_rejected",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Rejected",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategySignalListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Strategy Signals",
        "tags": [
          "strategy-runtime"
        ]
      }
    },
...
    "/api/v1/walkforward/results": {
      "get": {
        "description": "List walk-forward validation results for the current user.",
        "operationId": "list_walkforward_results_api_v1_walkforward_results_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalkForwardListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Walkforward Results",
        "tags": [
          "walk-forward"
        ]
      }
    },
    "/api/v1/walkforward/run": {
      "post": {
        "description": "Run walk-forward validation (IS backtest + OOS aggregate).",
        "operationId": "run_walkforward_api_v1_walkforward_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalkForwardRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalkForwardRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Walkforward",
        "tags": [
          "walk-forward"
        ]
      }
    },
    "/api/v1/walkforward/{run_id}": {
      "get": {
        "description": "Get one walk-forward validation run with IS/OOS report.",
        "operationId": "get_walkforward_api_v1_walkforward__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalkForwardRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Walkforward",
        "tags": [
          "walk-forward"
        ]
      }
    }
  }
}